core: split all patches into more descriptive files
