+++ BlasII.Randomizer/Patches/PopupMessagePatches.cs
@@ -1,10 +1,7 @@
 ﻿using BlasII.ModdingAPI;
 using HarmonyLib;
-using Il2CppI2.Loc;
 using Il2CppTGK.Game.Components.UI;
 using Il2CppTGK.Game.PopupMessages;
-using UnityEngine;
-using UnityEngine.UI;
 
 namespace BlasII.Randomizer.Patches;
 
@@ -12,7 +9,7 @@ namespace BlasII.Randomizer.Patches;
 /// Change display message when interacting with certain objects
 /// </summary>
 [HarmonyPatch(typeof(PopupMessageLogic), nameof(PopupMessageLogic.ShowMessageAndWait))]
-class Popup_Show_Patch
+class PopupMessageLogic_ShowMessageAndWait_Patch
 {
     public static void Postfix(PopupMessageLogic __instance, PopupMessage message)
     {
@@ -44,35 +41,5 @@ class Popup_Show_Patch
             __instance.textCtrl.SetText(text);
             return;
         }
-    }
-}
-
-/// <summary>
-/// When opening the map, replace the cherub count with collected items count
-/// </summary>
-[HarmonyPatch(typeof(MapWindowLogic), nameof(MapWindowLogic.UpdateUIElements))]
-class Map_Update_Patch
-{
-    public static void Postfix(MapWindowLogic __instance)
-    {
-        Transform cherubHolder = __instance.transform.Find("Background/UpperZone/PercentMapAndCherubs/cherubs");
-        if (cherubHolder == null)
-            return;
-
-        int currentItems = Main.Randomizer.ItemHandler.CollectedLocations.Count;
-        int totalItems = Main.Randomizer.ItemHandler.MappedItems.Count;
-
-        var leftText = cherubHolder.GetChild(0).GetChild(0).GetComponent<UIPixelTextWithShadow>();
-        leftText.SetText(currentItems.ToString());
-        Object.Destroy(leftText.GetComponent<Localize>());
-
-        var rightText = cherubHolder.GetChild(0).GetChild(1).GetComponent<UIPixelTextWithShadow>();
-        rightText.SetText("/" + totalItems);
-        Object.Destroy(rightText.GetComponent<Localize>());
-
-        var image = cherubHolder.GetChild(1).Cast<RectTransform>();
-        image.anchoredPosition = new Vector2(30, 0);
-        image.sizeDelta = new Vector2(80, 80);
-        image.GetComponent<Image>().sprite = Main.Randomizer.Data.GetImage(DataStorage.ImageType.Chest);
     }
 }
